PCLOUDS-1602: SFM refactor - extract logs sfm
